Rename "fd" to "io".
(Naming things, right?)

This module is about "read" and "write" and other things you can do on
handles which represent streams of data -- although even there, we're
already in trouble because in some contexts, "stream" implies a
sequenced and reliable bytestream, and here we also want to talk about
datagrams and potentially unsequenced and/or unreliable data. So what
should we call it?

"fd" isn't ideal, because even in POSIX "file descriptors" are about
more than just files. "socket" and "channel" and "port" all mean
specific things in related systems.

"io" isn't perfect either, because reading and writing to stream-like
(or datagram-like) data isn't the only way to do I/O. However, it is a
pretty fundamental way, and I think it has the potential to be a basic
primitive building block that a lot of other libraries can build on. So
I propose we go with "io".

Opinions and bikeshed paintbrushes welcome!